style: clear the clippy diagnostics this branch introduced (#26)
* chore: resolve branch-only Clippy findings

Clean up the twelve Clippy diagnostics introduced on this branch while leaving the warnings inherited from `master` unchanged.

Apply behavior-preserving fixes for redundant range checks, identity casts, collapsible matches, constant assertions, thread-local const initialization, and test-module placement. Convert the compile-time invariant to a `const` assertion and move the FFI test module to the end of the file without changing its contents.

Keep the four `too_many_arguments` cases explicitly allowed because they represent the existing public H1–H4 range API and changing their shape would affect C bindings and external consumers. Document the actual reason at each site.

Leave dependency-originated deprecation warnings for `base64` and `ring` to a separate change, since updating those paths may affect encoding behavior and constant-time comparison semantics.

Verify formatting, per-feature checks, Clippy, and tests across the complete feature matrix. Clippy diagnostics are reduced to the eleven warnings already present on `master`.

+++ boringtun/src/noise/session.rs
@@ -205,7 +205,7 @@ impl Session {
             panic!("The destination buffer is too small");
         }
 
-        let sending_key_counter = self.sending_key_counter.fetch_add(1, Ordering::Relaxed) as u64;
+        let sending_key_counter = self.sending_key_counter.fetch_add(1, Ordering::Relaxed);
 
         let (message_type, rest) = dst.split_at_mut(4);
         let (receiver_index, rest) = rest.split_at_mut(4);
